quick nav fix
